pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mistralai>=1.9.11",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
]
